app: demonstrate data flow from controller to view with fake DTO
DTO here is a Tuple3. Done out of laziness. And only on the previously
used list of users.
Should also be done for the group.
Maybe even go further - we know that username of an user is unique.
Don't even sent the user id, only username. Controller can figure out
user id with the username. Not sure if this else falls under the
category of "Confidentiality loss occuring".
This also allows for a better compartmentalization, as the view
absolutely does not need to know everything about a user.