app: models: add Group & User as DB entities
This commit consists of many changes:

* Changing our models to DB entities,
* Simplification of object querying due to `Model/Finder` concept of `EBean',
* Propagating change due to `Id` type migration from int to Long,
* Adding default set of users and must-have groups through the evolution
concept

Due to the fact that we want to move forward, I took it on myself to
implement the DB integration. This will most likely be bugged in a few
places, because I have not come around to implement tests which actually
check if the ORM mapping works correctly in all cases.

Considering we need to talk about our session management including
threat analysis, architecture analysis and other security policies -
this week - I wanted to remove most of the legwork as we have enough to
do anyway.

While the commit seems clean, it is not clean at all. The API of `Group`
and `User` differs by name, as well as some methods seem redundant
considering the `Finder` concept already allows a direct search with id
for an object. These are things I would like seen to be improved.