app: build: add jbcrypt to library dependencies
Due to the requirement that we cannot/should not use a Play-external
authentication library or a plugin, we cannot use Play Silhouette. Using
a library however, which only provides the Bcrypt algorithm as-is should
be fine.

I also checked that this library is compliant with what Julius said on
Wednesday, 10.10.18. No need to save the salt explicitly. API seems easy
enough, see [0].

[0] http://rny.io/playframework/bcrypt/2013/10/22/better-password-hashing-in-play-2.html