app: models: UserSession: remove cascadetype annotation
To explain:
Through this cascade type deleting a session also issued a delete on the
user. This is super wrong and only throws SQLException, as it should, as
it breaks the referential integrity constraint, i.e. deleting a user but
not deleting all other references of it in the DB, e.g. the groups he
is the owner of.