app: views: implement button helper function
Mostly used as a abstraction around Bootstrap, so all buttons we use in
our templates look the same.
I tried to map most of the useful properties of Bootstrap to enums. In
case something is missing, or you do not know which enum does what,
please consult the Bootstrap documentation [0]

[0]: https://getbootstrap.com/docs/4.0/components/buttons/