app: views: enable a owner/admin to delete a group from the view
To be honest, I do not like the whole "let's use a table to align
things" method. This should actually be a grid based on CSS flexbox
using a few <divs> and the appropriate CSS classes.
Therefore, I will build something based on that.

That will also hopefully remove the whole problem with HTML gets
displayed in tables.