app: models: delete user also deletes all his sessions
This is okay for now, but a user which gets deleted, has no idea about
this. His next request to our page, if he is logged in, will just
forward him to /login. As well as he has no error message on /login,
that his username does not exist anymore.