docs: security-policies: first draft of security policies document
It's not quite finished and there are a few things I want to talk about,
which is why some things are marked in red.
Furthermore, and this is just written down here so I do not forget, the
example Prof. Peine provided us with in exercise 5 has written down
specific extra policies with regards to the CIA triad. That is something
we kinda need to talk about together.